[CL-2438] More generalization
