+++ front/app/containers/Admin/projects/project/nativeSurvey/index.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useState } from 'react';
+import React, { Fragment, lazy, useState } from 'react';
 import { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'utils/cl-intl';
 import { useParams, useLocation } from 'react-router-dom';
@@ -8,6 +8,7 @@ import { Box, Title, Text } from '@citizenlab/cl2-component-library';
 import FormActions from 'components/FormBuilder/components/FormActions';
 import FormResults from 'components/FormBuilder/components/FormResults';
 import Button from 'components/UI/Button';
+const FormBuilder = lazy(() => import('components/FormBuilder/edit'));
 
 // i18n
 import messages from 'components/FormBuilder/messages';
@@ -26,6 +27,8 @@ import { colors } from 'utils/styleUtils';
 
 // Services
 import { downloadSurveyResults } from 'services/formCustomFields';
+import { getUpdatedConfiguration, nativeSurveyConfig } from './utils';
+import useFormCustomFields from 'hooks/useFormCustomFields';
 
 const Forms = ({ intl: { formatMessage } }: WrappedComponentProps) => {
   const { projectId } = useParams() as { projectId: string };
@@ -120,6 +123,27 @@ const Forms = ({ intl: { formatMessage } }: WrappedComponentProps) => {
         )}
       </Box>
     </>
+  );
+};
+
+export const SurveyFormBuilder = () => {
+  const { projectId, phaseId } = useParams() as {
+    projectId: string;
+    phaseId?: string;
+  };
+
+  const formCustomFields = useFormCustomFields({
+    projectId,
+    phaseId,
+  });
+
+  return (
+    <FormBuilder
+      builderConfig={getUpdatedConfiguration(
+        nativeSurveyConfig,
+        formCustomFields
+      )}
+    />
   );
 };
